Fixed lint warnings in I2C DC DC and Control and Status register source files

+++ PSF/Demo/PSF_EVB_Source/PowerDelivery/PDSource_App.c
@@ -95,7 +95,7 @@ UINT8 PDStack_Events(UINT8 u8PortNum, UINT8 u8PDEvent)
         {
             UPD_GPIOEnableDisable(u8PortNum,(UINT8)eUPD_PIO2, UPD_DISABLE_GPIO);
             #if (TRUE == INCLUDE_POWER_BALANCING)
-            PB_HandlePPMEvents (u8PortNum, eMCHP_PSF_TYPEC_DETACH_EVENT); 
+            PB_HandlePPMEvents (u8PortNum, (UINT8)eMCHP_PSF_TYPEC_DETACH_EVENT); 
             #endif 
             break;
         }
@@ -106,7 +106,7 @@ UINT8 PDStack_Events(UINT8 u8PortNum, UINT8 u8PDEvent)
             UPD_GPIOSetBufferType(u8PortNum,(UINT8)eUPD_PIO2,UPD_GPIO_SETBUF_PUSHPULL);
             UPD_GPIOSetClearOutput(u8PortNum,(UINT8)eUPD_PIO2,UPD_GPIO_CLEAR);
             #if (TRUE == INCLUDE_POWER_BALANCING)
-            PB_HandlePPMEvents (u8PortNum, eMCHP_PSF_TYPEC_CC1_ATTACH); 
+            PB_HandlePPMEvents (u8PortNum, (UINT8)eMCHP_PSF_TYPEC_CC1_ATTACH); 
             #endif 
             break;
         }
@@ -117,7 +117,7 @@ UINT8 PDStack_Events(UINT8 u8PortNum, UINT8 u8PDEvent)
             UPD_GPIOSetBufferType(u8PortNum,(UINT8)eUPD_PIO2,UPD_GPIO_SETBUF_PUSHPULL);
             UPD_GPIOSetClearOutput(u8PortNum,(UINT8)eUPD_PIO2,UPD_GPIO_SET);
             #if (TRUE == INCLUDE_POWER_BALANCING)
-            PB_HandlePPMEvents (u8PortNum, eMCHP_PSF_TYPEC_CC2_ATTACH);
+            PB_HandlePPMEvents (u8PortNum, (UINT8)eMCHP_PSF_TYPEC_CC2_ATTACH);
             #endif 
             break;
         }
@@ -145,7 +145,7 @@ UINT8 PDStack_Events(UINT8 u8PortNum, UINT8 u8PDEvent)
         case eMCHP_PSF_PD_CONTRACT_NEGOTIATED: 
         {
             #if (TRUE == INCLUDE_POWER_BALANCING)
-            PB_HandlePPMEvents (u8PortNum, eMCHP_PSF_PD_CONTRACT_NEGOTIATED);
+            PB_HandlePPMEvents (u8PortNum, (UINT8)eMCHP_PSF_PD_CONTRACT_NEGOTIATED);
             #endif 
             break; 
         }
@@ -153,7 +153,7 @@ UINT8 PDStack_Events(UINT8 u8PortNum, UINT8 u8PDEvent)
         case eMCHP_PSF_GET_SINK_CAPS_NOT_RCVD: 
         {
             #if (TRUE == INCLUDE_POWER_BALANCING)
-            PB_HandlePPMEvents (u8PortNum, eMCHP_PSF_GET_SINK_CAPS_NOT_RCVD);
+            PB_HandlePPMEvents (u8PortNum, (UINT8)eMCHP_PSF_GET_SINK_CAPS_NOT_RCVD);
             #endif  
             break; 
         }  
@@ -161,7 +161,7 @@ UINT8 PDStack_Events(UINT8 u8PortNum, UINT8 u8PDEvent)
         case eMCHP_PSF_GET_SNK_CAPS_RCVD:
         {
             #if (TRUE == INCLUDE_POWER_BALANCING)
-            PB_HandlePPMEvents (u8PortNum, eMCHP_PSF_GET_SNK_CAPS_RCVD);
+            PB_HandlePPMEvents (u8PortNum, (UINT8)eMCHP_PSF_GET_SNK_CAPS_RCVD);
             #endif  
             break;            
         }
